WIFI-638: Reported Noise floor looks inaccurate
Should have noise floor sent as a signed value, is currently coming as an uint32
I believe the AP actually translates the signed value to that.
ON channel reports more than just the channel in use on the radio which is making a mess as well, I have added a filter on the OSGW as a workaround which will prevent OFF channel values being used during calculation of channel info for on channel reporting. This may become unnecessary when the AP fix comes in but for now might clear things up a bit. After we change to a signed value, combined with correct channel survey reporting, we should have a better value.