WIFI-1421: OSGW refactor OvsdbDao
Break OvsdbDao into a number of task oriented device adaptor classes classes.
OvsdbDao public methods now call these adapters, the callers to OvsdbDao
will not have to change.
Original methods and flow retained in this first step of refactoring, so
that the code is still following the same logic.
Going forward the actual inner workings of the smaller adapter classes
should be refactored as well.