WIFI-1514: AP-NOS Network: Client Devices: Client device status not updated correctly.
Adding logs to aide debugging.
Ensure associationStatus is set properly for ClientSessions.
IP address is set correctly initially, for ClientSessions, it gets set
incorrectly due to the incorrect ByteString set by the AP in the
ClientIpEvent. This overwrites the previous good value, for example,
10.0.0.250, with a nonsense value, i.e. 104.19.189.0
AP needs to investigate this part of the issue.